DSS SKILL/requirements.txt
Faker
speedtest-cli
keyboard